Namespaces in cbuf
With this change, cbuf has first class namespaces.
Namespaces will be merged, they do not need to appear contiguously.
Namespaces can be referred to, namespace A refers to B.
Namespaces is now part of the integral name of the message, this has
implications for logtool and filtering, but it is consistent.